extract-kv-cache-dtype-move: Move configure_kv_cache_dtype + TORCH_DTYPE_TO_KV_CACHE_STR to mem_cache.kv_cache_dtype (cut+paste)
Refactor chain ID: extract-kv-cache-dtype-move
